Moved Rarity & Chance to the api module
