+++ tradingcards-plugin/src/main/java/net/tinetwork/tradingcards/tradingcardsplugin/config/settings/RaritiesConfig.java
@@ -1,5 +1,6 @@
 package net.tinetwork.tradingcards.tradingcardsplugin.config.settings;
 
+import net.tinetwork.tradingcards.api.model.Rarity;
 import net.tinetwork.tradingcards.tradingcardsplugin.TradingCards;
 import net.tinetwork.tradingcards.tradingcardsplugin.core.SimpleConfigurate;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -25,35 +26,7 @@ public class RaritiesConfig extends SimpleConfigurate {
         return rootNode.node(id).get(Rarity.class);
     }
 
-    public static class Rarity {
-        private String name;
-        private String displayName;
-        private String defaultColor;
-        private List<String> rewards;
-
-        public Rarity(String name, String displayName, String defaultColor, List<String> rewards) {
-            this.name = name;
-            this.displayName = displayName;
-            this.defaultColor = defaultColor;
-            this.rewards = rewards;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public String getDisplayName() {
-            return displayName;
-        }
-
-        public String getDefaultColor() {
-            return defaultColor;
-        }
-
-        public List<String> getRewards() {
-            return rewards;
-        }
-    }
+  
 
     public static final class RaritySerializer implements TypeSerializer<Rarity> {
         public static final RaritySerializer INSTANCE = new RaritySerializer();
